+++ config/tracking_providers.php
@@ -1,9 +1,11 @@
 <?php
 
+use App\Services\Fleetboard\FleetboardService;
 use App\Services\Mixtelematics\MixIntegrateService;
 
 return [
     'services' => [
-        'powerfleet' => MixIntegrateService::class
+        'powerfleet' => MixIntegrateService::class,
+        'fleetboard' => FleetboardService::class,
     ],
 ];